Split launcher and reaper
