+++ daemons/Makefile
@@ -15,7 +15,8 @@ PARALLEL_AWS_DAEMONS := \
 	dss-s3-copy-write-metadata-sfn \
 	dss-visitation \
 	dss-scalability-test \
-	dss-sfn \
+	dss-dlq-reaper \
+	dss-sfn-launcher \
 
 PARALLEL_GCP_DAEMONS := \
 	dss-gs-event-relay
@@ -25,7 +26,7 @@ deploy-serial: $(SERIAL_AWS_DAEMONS) $(SERIAL_GCP_DAEMONS)
 deploy-parallel: $(PARALLEL_AWS_DAEMONS) $(PARALLEL_GCP_DAEMONS)
 
 $(SERIAL_AWS_DAEMONS) $(PARALLEL_AWS_DAEMONS) scheduled-ci-build:
-	@if [[ $@ == "dss-sfn" ]]; then \
+	@if [[ $@ == "dss-dlq-reaper" || $@ == "dss-sfn-launcher" ]]; then \
    	        $(DSS_HOME)/scripts/deploy_dss_sfn_queue.py; \
         fi
 	./package_daemon.sh $@
@@ -36,7 +37,7 @@ $(SERIAL_AWS_DAEMONS) $(PARALLEL_AWS_DAEMONS) scheduled-ci-build:
             ../tests/daemons/sample_s3_bundle_created_event.json.template \
             ../tests/daemons/a47b90b2-0967-4fbf-87bc-c6c12db3fedf.2017-07-12T055120.037644Z; \
         fi
-	@if [[ $@ == "dss-s3-copy-sfn" || %@ == "dss-s3-copy-write-metadata-sfn" || $@ == "dss-checkout" || $@ == "dss-scalability-test" ]]; then \
+	@if [[ $@ == "dss-s3-copy-sfn" || %@ == "dss-s3-copy-write-metadata-sfn" || $@ == "dss-checkout" ]]; then \
             $(DSS_HOME)/scripts/deploy_checkout_lifecycle.py; \
         fi
