+++ daemons/dss-sfn-launcher/requirements.txt
@@ -1,0 +1,55 @@
+# You should not edit this file directly.  Instead, you should edit requirements.txt.in.
+asn1crypto==0.23.0
+azure-common==1.1.8
+azure-nspkg==2.0.0
+azure-storage==0.36.0
+boto3==1.6.0
+botocore==1.9.0
+cachetools==2.0.1
+certifi==2018.1.18
+cffi==1.11.2
+chardet==3.0.4
+click==6.7
+clickclick==1.2.2
+cloud-blobstore==2.1.0
+connexion==1.1.15
+cryptography==2.1.4
+docutils==0.14
+elasticsearch==5.5.1
+elasticsearch-dsl==5.3.0
+fasteners==0.14.1
+Flask==0.12.2
+google-api-core==0.1.1
+google-apitools==0.5.21
+google-auth==1.2.1
+google-cloud-core==0.28.0
+google-cloud-storage==1.8.0
+google-resumable-media==0.3.1
+googleapis-common-protos==1.5.3
+httplib2==0.11.0
+idna==2.6
+inflection==0.3.1
+iso8601==0.1.12
+itsdangerous==0.24
+Jinja2==2.10
+jmespath==0.9.3
+jsonschema==2.6.0
+MarkupSafe==1.0
+monotonic==1.4
+nestedcontext==0.0.4
+oauth2client==4.1.2
+protobuf==3.5.0.post1
+pyasn1==0.4.2
+pyasn1-modules==0.2.1
+pycparser==2.18
+python-dateutil==2.6.1
+PyYAML==3.12
+requests==2.18.4
+requests-aws4auth==0.9
+requests-http-signature==0.0.3
+rsa==3.4.2
+s3transfer==0.1.11
+six==1.11.0
+swagger-spec-validator==2.1.0
+urllib3==1.22
+Werkzeug==0.12.2
